+++ src/pfddm.cpp
@@ -11,9 +11,9 @@
 
 //' Distribution of Ratcliff Diffusion Decision Model
 //'
-//' Distribution function for the Ratcliff diffusion decision model (DDM) with
-//' following parameters: \code{a} (threshold separation), \code{v} (drift
-//' rate), \code{t0} (non-decision time/response time constant), \code{w}
+//' Density function for the Ratcliff diffusion decision model (DDM) with
+//' following parameters: \code{v} (drift rate), \code{a} (threshold
+//' separation), \code{t0} (non-decision time/response time constant), \code{w}
 //' (relative starting point), \code{sv} (inter-trial variability of drift), and
 //' \code{sigma} (diffusion coefficient of underlying Wiener process).
 //'
@@ -40,10 +40,6 @@
 //'         TRUE \ifelse{html}{\out{&#8594;}}{\eqn{\to}} "upper");
 //'     }
 //'
-//' @param a Threshold separation. Amount of information that is considered for
-//'   a decision. Large values indicate a conservative decisional style. Allowed
-//'   range: \eqn{0 <} \code{a}. Typical range: \eqn{0.5 <} \code{a} \eqn{< 2}.
-//'
 //' @param v Drift rate. Average slope of the information accumulation process.
 //'   The drift gives information about the speed and direction of the
 //'   accumulation of information. Large (absolute) values of drift indicate a
@@ -52,6 +48,10 @@
 //'   value indicates that the received information supports the response linked
 //'   to the lower threshold. Allowed range: \code{v} is a real number. Typical
 //'   range: \eqn{-5 <} \code{v} \eqn{< 5}.
+//'
+//' @param a Threshold separation. Amount of information that is considered for
+//'   a decision. Large values indicate a conservative decisional style. Allowed
+//'   range: \eqn{0 <} \code{a}. Typical range: \eqn{0.5 <} \code{a} \eqn{< 2}.
 //'
 //' @param t0 Non-decision time or response time constant (in seconds). Lower
 //'   bound for the duration of all non-decisional processes (encoding and
@@ -99,12 +99,12 @@
 //' @details
 //'
 //' All of the model inputs and parameters (\code{rt}, \code{response},
-//' \code{a}, \code{v}, \code{t0}, \code{w}, \code{sv}, \code{sigma}) can be
+//' \code{v}, \code{a}, \code{t0}, \code{w}, \code{sv}, \code{sigma}) can be
 //' input as a single value or as a vector of values. If input as a vector of
 //' values, then the standard \code{R} input wrapping will occur.
 //'
 //' \code{sigma} - The default value of this parameter is \code{1} because it
-//' only scales the parameters \code{a}, \code{v}, and \code{sv}, as shown
+//' only scales the parameters \code{v}, \code{a}, and \code{sv}, as shown
 //' above. However, other formulations of the DDM may set \code{sigma = 0.1}
 //' (see Ratcliff (1978), the fourth reference), so care must be taken when
 //' comparing the results of different formulations.
@@ -142,8 +142,8 @@
 // [[Rcpp::export]]
 NumericVector pfddm(const NumericVector& rt,
                     const SEXP& response,
-                    const NumericVector& a,
                     const NumericVector& v,
+                    const NumericVector& a,
                     const NumericVector& t0,
                     const NumericVector& w = 0.5,
                     const NumericVector& sv = 0.0,
@@ -160,14 +160,14 @@ NumericVector pfddm(const NumericVector& rt,
   
   // determine lengths of parameter inputs, except response
   int Nrt  = rt.length();
-  int Na   = a.length();
   int Nv   = v.length();
+  int Na   = a.length();
   int Nt0  = t0.length();
   int Nw   = w.length();
   int Nsv  = sv.length();
   int Nsig = sigma.length();
   int Nerr = err_tol.length();
-  int Nmax = max({Nrt, Na, Nv, Nt0, Nw, Nsv, Nsig, Nerr}); // include Nres later
+  int Nmax = max({Nrt, Nv, Na, Nt0, Nw, Nsv, Nsig, Nerr}); // include Nres later
   int Nres;
 
 
@@ -176,16 +176,16 @@ NumericVector pfddm(const NumericVector& rt,
 
 
   // check for invalid inputs, invalid inputs get marked in the vector `out`
-  if (!parameter_check(Nrt, Nres, Na, Nv, Nt0, Nw, Nsv, Nsig, Nerr, Nmax,
-                       rt, response, a, v, t0, w, sv, sigma, err_tol,
+  if (!parameter_check(Nrt, Nres, Nv, Na, Nt0, Nw, Nsv, Nsig, Nerr, Nmax,
+                       rt, response, v, a, t0, w, sv, sigma, err_tol,
                        out, rt0)) {
     NumericVector empty_out(0);
     return empty_out;
   }
 
   // loop through all inputs, the vector `out` gets updated
-  calculate_cdf(Nrt, Na, Nv, Nt0, Nw, Nsv, Nsig, Nerr,
-                Nmax, rt, a, v, t0, w, sv, sigma,
+  calculate_cdf(Nrt, Nv, Na, Nt0, Nw, Nsv, Nsig, Nerr,
+                Nmax, rt, v, a, t0, w, sv, sigma,
                 err_tol, out, rt0, disf);
 
   return Rcpp::wrap(out);